1) Supports partitioning abstraction(work in progress) 2) Worklist consider a set of jobs for the "next job" operation. 3) JSON load/store stuffs does not work in this version. 4) The analyzer console also does not show the abstract states properly.

+++ src/main/scala/kr/ac/kaist/safe/cfg_builder/DotWriter.scala
@@ -17,7 +17,7 @@ import kr.ac.kaist.safe.nodes.cfg._
 import kr.ac.kaist.safe.LINE_SEP
 
 object DotWriter {
-  private type OrderMap = Map[CFGBlock, Int]
+  private type OrderMap = Map[CFGBlock, Long]
   val CurBlockShape: String = "shape=record, style=bold, fontsize=11"
   val NormalBlockShape: String = "shape=record, fontsize=11"
   val NormalEdgeStyle: String = "style=solid"
